pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rv-ai-skills-2"
version = "0.1.0"
description = "CSV to report pipeline with schema-first report types and OpenAI-ready orchestration."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.2.0",
  "pydantic>=2.8.0",
  "PyYAML>=6.0.1",
  "jsonschema>=4.22.0",
  "jinja2>=3.1.4",
  "typer>=0.12.3",
  "rich>=13.7.1",
  "flask>=3.0.3",
  "playwright>=1.50.0",
  "xhtml2pdf>=0.2.15",
]

[project.optional-dependencies]
openai = ["openai>=1.40.0"]
dev = ["pytest>=8.2.0", "httpx>=0.27.0", "ruff>=0.5.0"]

[project.scripts]
rv-reporter = "rv_reporter.cli:app"
rv-reporter-web = "rv_reporter.web:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
